+++ CHANGELOG.md
@@ -6,9 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Facebook color palette.
 - Whatsapp color palette.
-- Flutter greyish color palette.
+- Flutter color palette.
+
+## [0.0.3] - 2021-04-06
+### Added
+- Facebook color palette. [#3](https://github.com/rafamizes/flutter_brand_palettes/issues/3)
 
 ## [0.0.2] - 2021-04-05
 ### Added